stop using autoreconfHook
Currently some of the Ruby versions fail to build. This is because the
`configuration` file that autoreconf generates is incorrect. Probably
because of configuration.ac for some of the Ruby versions.

It seems that most Ruby verions nowadays build fine with their
distributed `configuration` file, so there is no need anymore to reconf
these.